requirements.txt
annotated-types==0.7.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
deep-sort-realtime==1.3.2
exceptiongroup==1.2.2
filelock==3.18.0
flake8==7.2.0
flasgger==0.9.7.1
Flask==3.1.0
fonttools==4.58.0
fsspec==2025.3.2
git-filter-repo==2.47.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mistune==3.1.3
mpmath==1.3.0
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycodestyle==2.13.0
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
types-requests==2.32.0.20250515
tzdata==2025.2
ultralytics==8.3.136
ultralytics-thop==2.0.14
urllib3==2.4.0
uuid==1.30
Werkzeug==3.1.3
